requirements.txt
pkg-resources==0.0.0
PyYAML==3.11
requests==2.10.0
six==1.10.0
slackclient==1.0.1
websocket-client==0.37.0
